api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
paypalrestsdk==1.13.1
flask-mail==0.9.1
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0
werkzeug==2.0.1
jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
markupsafe==2.0.1
PyJWT==2.3.0
dnspython==2.1.0 